requirements.txt
spotipy~=2.19.0
python-dotenv~=0.19.2
telethon~=1.24.0